dom/tests/mochitest/notification/test_bug893501.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>bug893501 - crash test</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="notification_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<div id="foo"></div>
</div>
<pre id="test">
<script type="application/javascript"><!--
if (is_feature_enabled()) {
  SimpleTest.waitForExplicitFinish();

  function showNotifications() {
    ok(navigator.mozNotification, "test for notification.");

    var div = document.getElementById('foo');
    var frame = document.createElement('iframe');
    div.appendChild(frame);

    var frameNavigator = frame.contentWindow.navigator;
    frame.addEventListener("load", n, false);
    frame.src = "data:text/html,hello world";

    function n()
    {
      ok(frameNavigator.mozNotification, "Still alive!");
      SimpleTest.finish();
    }
  }

  setup_notifications(true, false, showNotifications);
} else {
  ok(true, "Desktop notifications not enabled.");
}
// -->
</script>
</pre>
</body>
</html>
